+++ lydian/cogs/debug.py
@@ -3,7 +3,7 @@ from discord import Embed
 from discord.ext import commands
 from loguru import logger
 
-from lydian.cogs.util import embed_error, embed_info, embed_ok, embed_warn
+from lydian.cogs.util import confirm, embed_error, embed_info, embed_ok, embed_warn
 from lydian.config import config
 
 
@@ -35,11 +35,26 @@ class DebugCog(commands.Cog):
         await ctx.send(embed=embed_error('Error', 'Something failed.'))
 
     @commands.command(checks=[debug_enabled])
+    async def promptyn(self, ctx: commands.Context, prompt_timeout: float | None = None) -> None:
+        """Sends a message with button components.
+
+        Accepts an optional argument to override the timeout in seconds.
+        """
+        response = await confirm(ctx, embed_info('Continue?'), prompt_timeout=prompt_timeout)
+        match response:
+            case True:
+                await ctx.send(embed=embed_info('Confirmed.'))
+            case False:
+                await ctx.send(embed=embed_info('Cancelled.'))
+            case _:
+                await ctx.send(embed=embed_info('Prompt timed out.'))
+
+    @commands.command(checks=[debug_enabled])
     async def fail(self, _ctx: commands.Context) -> None:
         """Raises a ``ValueError``."""
         logger.debug('Raising ValueError...')
         raise ValueError('Testing, testing!')
-        logger.debug('Raised ValueError; this message should not be visible')
+        logger.debug('Raised ValueError; this message should not be visible!')
 
     @commands.command(checks=[debug_enabled])
     async def bigembed(self, ctx: commands.Context) -> None: